Move: container and presets files
