+++ test/types.js
@@ -1,4 +1,4 @@
 /**
  * @typedef TestCreateContainer
- * @type {(option: ContainerOptions, run: boolean) => import('../src/docker/containers/container')} - the modified function of CreateContainer.
+ * @type {(option: ContainerOptions, run: boolean) => import('../src/docker/container')} - the modified function of CreateContainer.
  */
